Added result when domain isn't allowed
Added some code to deal with the situation when an email address domain isn't one of the allowed. Email is moved to the 'faulted' folder and script returns to emailrelay.exe with the exit code 65, it seemed roughly appropriate. Don't know enough about email to know if that is accurate.

Also change the constants that deal with folder locations to be based off of a baseDir variable, to make it easier to change around.