Fixed logic, added usernames to log
* Fixed logic loop

This required added in a bool and switching it when we find a match. I'm sure there's a better way, I was in a hurry.

* Added usernames to log file

Added an extra catch in the regexp to grab the username (everything before the @) and add it to the log as well. This way I know who in particular was doing it.